requirements.txt
anthropic
python-telegram-bot
psycopg2-binary
python-dotenv